transport: allow for custom non-QUIC packet detection
Existing QUIC packed detection is quite rudimentary. In some advanced
cases, the user might want to have more control over exact detection of
QUIC and Non-QUIC packets.

This commit simply allows a function to be defined on the Transport
which overrides this simple handling. Should no such function be
defined, the processing falls back to the previous method.